+++ src/main/java/uk/ac/leeds/ccg/grids/d2/stats/Grids_StatsNotUpdatedInt.java
@@ -18,7 +18,6 @@ package uk.ac.leeds.ccg.grids.d2.stats;
 import java.io.IOException;
 import java.math.BigDecimal;
 import uk.ac.leeds.ccg.grids.core.Grids_Environment;
-import java.math.BigInteger;
 
 /**
  * Statistic fields are not kept up to date as the underlying data is changed.
@@ -94,7 +93,7 @@ public class Grids_StatsNotUpdatedInt extends Grids_StatsInt {
      * @throws java.lang.ClassNotFoundException If encountered.
      */
     @Override
-    public BigInteger getN() throws IOException, Exception, ClassNotFoundException {
+    public long getN() throws IOException, Exception, ClassNotFoundException {
         update();
         return n;
     }
